pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kie-api"
version = "0.1.0"
description = "Reusable KIE.AI integration toolkit for model specs, normalization, validation, and task orchestration."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["kie", "ai", "sdk", "video-generation", "image-generation", "api-client"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27,<1.0",
  "Pillow>=10.0,<12.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
kie_api = [
  "py.typed",
  "resources/specs/models/*.yaml",
  "resources/prompt_profiles/*/metadata.yaml",
  "resources/prompt_profiles/*/prompt.md",
  "resources/pricing/*.yaml",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "smoke: optional live smoke tests that require explicit KIE env vars",
]
